move isolated geometry tests to connected buffer geometry
