+++ test/connectedBufferGeometry.js
@@ -1,13 +1,48 @@
-import { ConnectedBufferGeometry, STLLoader } from '..';
-import * as THREE from 'three';
+import { BufferGeometryAnalyzer, ConnectedBufferGeometry, STLLoader } from '..';
 import { expect } from 'chai';
 import fs from 'fs';
 
-describe("connectedBufferGeometry", function() {
-    it("Constructor", function() {
-        let stl = fs.readFileSync("test/DINOSAUR_JUMP.stl", {encoding: "ascii"});
-        let geometry = new STLLoader().parse(stl);
-        let x = new ConnectedBufferGeometry().fromBufferGeometry(geometry);
-        let y = x.isolatedBufferGeometries();
+describe("ConnectedBufferGeometry", function() {
+    describe("isolatedBufferGeometries", function() {
+        let testFile = function (filename, expectedGeometriesCount) {
+            let stl = fs.readFileSync("test/" + filename, {encoding: "ascii"});
+            let geometry = new STLLoader().parse(stl);
+            let connectedBufferGeometry = new ConnectedBufferGeometry().fromBufferGeometry(geometry);
+            let newGeometries = connectedBufferGeometry.isolatedBufferGeometries(geometry);
+            expect(newGeometries.length).to.equal(expectedGeometriesCount);
+        }
+
+        it("Simple tetrahedron", function() {
+            testFile("tetrahedron.stl", 1);
+        });
+
+        it("Split ruler with degenerate facets", function() {
+            testFile("lungo.stl", 2);
+        });
+
+        it("2 tetrahedrons that share a face", function() {
+            testFile("face_connected_tetrahedrons.stl", 2);
+        });
+
+        it("2 tetrahedrons that share an edge", function() {
+            testFile("edge_connected_tetrahedrons.stl", 2);
+        });
+
+        it("27 cubes in 3 by 3 by 3 formation", function() {
+            testFile("rubix.stl", 27);
+        });
+
+        it("27 cubes in 3 by 3 by 3 formation on an angle", function() {
+            testFile("twisted_rubix.stl", 27);
+        });
+
+        it("27 cubes in 3 by 3 by 3 formation with facets in lightly shuffled order", function() {
+            testFile("shuffled_rubix.stl", 27);
+        });
+
+        it("Big object: Dinosaur Jump", function() {
+            this.timeout(10000);
+            testFile("DINOSAUR_JUMP.stl", 1);
+        });
     });
 });
